requirements.txt
apturl==0.5.2
asgiref==3.4.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
bootstrap4==0.1.0
Brlapi==0.7.0
certifi==2021.10.8
chardet==3.0.4
cloudinary==1.26.0
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.7
django-bootstrap-v5==1.0.5
django-bootstrap4==21.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-guardian==2.4.0
django-heroku==0.3.1
django-registration==1.0
django-registration-redux==2.9
django-tinymce==3.4.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-metadata==3.10.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pbr==5.8.0
pexpect==4.6.0
Pillow==8.3.2
protobuf==3.6.1
psycopg2==2.9.1
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.5
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
wadllib==1.3.3
whitenoise==5.3.0
xkit==0.0.0
zipp==3.6.0
